+++ docker/setup.sh
@@ -300,7 +300,7 @@ fi
 
 log "Enabling forwarding and loose reverse-path filtering."
 cat > /etc/sysctl.d/99-warp-egress-docker.conf <<SYSCTL
-net.ipv4.ip_forward=1
+net.ipv4.ip_forward=0
 net.ipv4.conf.all.rp_filter=2
 net.ipv4.conf.default.rp_filter=2
 SYSCTL